# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Paging:
    properties:
      Page:
        type: integer
      size:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  dto.ResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.ResponseMessage:
    properties:
      message:
        type: string
    type: object
  dto.ResponsePaging:
    properties:
      data: {}
      message:
        type: string
      paging:
        $ref: '#/definitions/dto.Paging'
    type: object
  dto.TopUpReq:
    properties:
      top up amount:
        type: integer
    type: object
  dto.TopUpupdate:
    properties:
      accepted status:
        type: boolean
      id:
        type: string
      information status:
        type: string
    type: object
...
info:
  contact: {}
  title: Laundry Apps
  version: "1.0"
paths:
...
  /topup:
    post:
      consumes:
      - application/json
      description: create new data topup
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data topup
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TopUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: new
      tags:
      - topup
  /topup/{id}:
    get:
      consumes:
      - application/json
      description: get data top up by id data
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get by id data
      tags:
      - topup
  /topup/confirm:
    get:
      consumes:
      - application/json
      description: get updated data topup
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get data updated
      tags:
      - topup
    put:
      consumes:
      - application/json
      description: confirm data topup
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data top up
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TopUpupdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: confirm
      tags:
      - topup
  /topup/list/{page}/{size}:
    get:
      consumes:
      - application/json
      description: get all data saldo
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page of pagination
        in: path
        name: page
        required: true
        type: integer
      - description: size of pagination
        in: path
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponsePaging'
      summary: get all
      tags:
      - topup
  /topup/upload:
    post:
      consumes:
      - multipart/form-data
      description: upload recipe
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: ID parameter
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: upload
      tags:
      - topup
  /topup/user:
    get:
      consumes:
      - application/json
      description: get data top up by user login
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get by user login
      tags:
      - topup
  /topup/user/{id}:
    get:
      consumes:
      - application/json
      description: get data top up by user id
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get by id user
      tags:
      - topup
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"